requirements.txt
flask==2.3.3
werkzeug==2.3.7
python-dotenv==1.0.0
openai==1.3.0
httpx==0.24.1 